Snippets for SQLite and CSV queries
Added basic snippets for querying SQLite and CSV databases with virtually identical interface.
Note, for CSV, the driver is set to the path without the file name and the table name is set to the file name. For SQLite, database is set to pathname and the table name is set to actual table name.