UseDisconnectedRecordset
Added UseDisconnectedRecordset to the IDbCommandFactory interface to be used in DbCommand. DbCommand object should not be responsible for establishing DbConnection either way. But it can still return either online or disconnected recordset using this flag. Is not used in either AutoDbCommand or DefaultDbCommand, but the relevant interface signatures required appropriate adjustment according to  IDbCommandFactory